Move StateNamespace and related to runners/core-java
