+++ runners/core-java/src/main/java/org/apache/beam/runners/core/StateNamespaces.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.util.state;
+package org.apache.beam.runners.core;
 
 import com.google.common.base.Splitter;
 import java.io.IOException;
