+++ runners/core-java/src/test/java/org/apache/beam/runners/core/TimerInternalsTest.java
@@ -29,8 +29,6 @@ import org.apache.beam.sdk.testing.CoderProperties;
 import org.apache.beam.sdk.transforms.windowing.GlobalWindow;
 import org.apache.beam.sdk.transforms.windowing.IntervalWindow;
 import org.apache.beam.sdk.util.TimeDomain;
-import org.apache.beam.sdk.util.state.StateNamespace;
-import org.apache.beam.sdk.util.state.StateNamespaces;
 import org.joda.time.Instant;
 import org.junit.Test;
 import org.junit.runner.RunWith;
